Refactor, focusing on .sample() and .monte_carlo_logsumexp() methods
